# entertainmentEmporiumBackend/schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Entertainment Emporium Backend API 
  description: A movies and tv shows rating API 
  contact:
    name: Francesca 
    email: passmoref@coventry.ac.uk
  license:
    name: Apache 2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
servers:
  - url: 'https://hexagonprofile-maximumalex-3000.codio-box.uk/api/v1'
    description: Development server
...
paths:
...
  /actors/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the actor
    get:
      tags:
        - Actors
      description:
        Get actor by id.
      responses:
        '200':
          description: View of a actor's record.
        '404':
          description: Actor was not found in the DB.
    put:
      tags:
        - Actors
      description:
        Update actor by ID.
      responses:
        '200':
          description: Update confirmation.
        '401':
          description: Only admin users can update actors.
        '404':
          description: Actor was not found in the DB.
        '403':
          description: Requester forbidden to update this resource.

    delete:
      tags:
        - Actors
      description:
        Delete actor by ID.
      responses:
        '200':
          description: Delete confirmation.
        '401':
          description: Only admin users can delete actors.
        '404':
          description: Actor was not found in the DB.
        '403':
          description: Requester forbidden to delete this resource
  /actors:
    summary: Represents an individual actor
    description: >
      This resource represents an individual actor in the system.
      Each actor is identified by a numeric `id`.
    post:
      tags:
        - Actors
      description:
        Create a new actor.
      requestBody:
        description: All new actor data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./actors.json#/definitions/actor
      responses:
        '201':
          description: Actor created
          content:
            application/json:
              schema:
                $ref: ./actors.json#/definitions/actorCreated
  /actors/{actors}:
    parameters:
      - in: path
        name: actors
        schema:
          type: string
        required: true
        description: A list of ids of actors e.g 1,2,3
    get:
      tags:
        - Actors
      description:
        Get multiple actors
      responses:
        '200':
          description: View of multiple actors.
        '404':
          description: Actors not found.
  /actors/search/{name}:
    parameters:
      - in: path
        name: name
        schema:
          type: string
        required: true
        description: Last name of an actor
    get:
      tags:
        - Actors
      description:
        Get an actor by last name
      responses:
        '200':
          description: View of actors with the last name searched.
        '404':
          description: Actors not found.
  /directors/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Numeric id of the director
    get:
      tags:
        - Directors
      description:
        Get director by id.
      responses:
        '200':
          description: View of a director's record.
        '404':
          description: Director was not found in the DB.
    put:
      tags:
        - Directors
      description:
        Update director by ID.
      responses:
        '200':
          description: Update confirmation
        '401':
          description: Only admin users can update directors.
        '404':
          description: Actor was not found in the DB.
        '403':
          description: Requester forbidden to update this resource.

    delete:
      tags:
        - Directors
      description:
        Delete director by ID.
      responses:
        '200':
          description: Delete confirmation.
        '401':
          description: Only admin users can delete directors.
        '404':
          description: Director was not found in the DB.
        '403':
          description: Requester forbidden to delete this resource
  /directors:
    summary: Represents an individual director
    description: >
      This resource represents an individual director in the system.
      Each director is identified by a numeric `id`.
    post:
      tags:
        - Directors
      description:
        Create a new director.
      requestBody:
        description: All new director data.
        required: true
        content:
          application/json:
            schema:
              $ref: ./directors.json#/definitions/director
      responses:
        '201':
          description: Actor created
          content:
            application/json:
              schema:
                $ref: ./directors.json#/definitions/directorCreated
  /directors/{directors}:
    parameters:
      - in: path
        name: directors
        schema:
          type: string
        required: true
        description: A list of ids of directors e.g 1,2,3
    get:
      tags:
        - Directors
      description:
        Get multiple directors
      responses:
        '200':
          description: View of multiple directors.
        '404':
          description: Directors not found.
  /directors/search/{name}:
    parameters:
      - in: path
        name: name
        schema:
          type: string
        required: true
        description: Last name of an director
    get:
      tags:
        - Directors
      description:
        Get a director by last name
      responses:
        '200':
          description: View of directors with the last name searched.
        '404':
          description: Directors not found.
...